requirements.txt
anthropic
python-dotenv==1.0.0
tqdm==4.66.1
jsonschema==4.20.0
rich==13.6.0
typer==0.9.0
pydrive==1.3.1
matplotlib==3.7.1
numpy==1.24.3
pandas==2.0.1
requests==2.31.0 